elbepack: separate submit/download logic from VM logic
elbepack/initvm(action).py currently couples the SOAP submit/download
logic (used by elbe initvm submit) with parts only relevant for the
actual VM initialization. To enable a future command that does not
depend on a VM, extract the parts relevant for both commands into
separate files.

This commit only moves code around. No functional change.
