+++ debian/python3-elbe-common.install
@@ -11,6 +11,7 @@ usr/lib/python3.*/*-packages/elbepack/commands/__init__.py
 usr/lib/python3.*/*-packages/elbepack/aptpkgutils.py
 usr/lib/python3.*/*-packages/elbepack/aptprogress.py
 usr/lib/python3.*/*-packages/elbepack/archivedir.py
+usr/lib/python3.*/*-packages/elbepack/buildsubmitaction.py
 usr/lib/python3.*/*-packages/elbepack/changelogxml.py
 usr/lib/python3.*/*-packages/elbepack/cli.py
 usr/lib/python3.*/*-packages/elbepack/config.py
@@ -37,6 +38,7 @@ usr/lib/python3.*/*-packages/elbepack/repodir.py
 usr/lib/python3.*/*-packages/elbepack/rfs.py
 usr/lib/python3.*/*-packages/elbepack/xmlpreprocess.py
 usr/lib/python3.*/*-packages/elbepack/shellhelper.py
+usr/lib/python3.*/*-packages/elbepack/soaphelper.py
 usr/lib/python3.*/*-packages/elbepack/spdx.py
 usr/lib/python3.*/*-packages/elbepack/templates.py
 usr/lib/python3.*/*-packages/elbepack/treeutils.py
